Capstone _Project_week1
As known, London is considered one of the most important, most expensive, and most visited cities around the globe. It has diverse cultures with various races and demographics. In short, London is a mix of everything and we can write uncountable pages describing its culture, history, and economy. Thus, in this project, I will address the issue of moving to the city of London as an ex-pat or as international students or any other reasons that make you move to London. I will consider some criteria in listing the best boroughs to live or to work or to study based on safety measurement, average rental rate. Besides, I will address the population rate in each borough for further measurement and analysis. Once again, my concentration is the safety and the renting rate. Furthermore, I did not point out about the transportation services or the entertainment venues since I do believe that London is a well-linked city in terms of transportation services, and with that, it will be easy to reach entertaining venues through various transportation services that the city provides to its people and visitors.